reformat benchmark and replace origin benchmark file
